requirements.txt
numpy
pandas
tqdm
faiss-cpu
psycopg
sentence-transformers
python-dotenv
